[unnecessary copy] Address copied return value
Summary:
This diff makes the checker to understand copied return values. When a copied value was passed as a return value and it was not modified, the checker did not report it as an unnecessary copy.

There were two issues:

1: When a return object was copy-constructed/assigned, it was not recognized by the checker because the LHS was a pointer in SIL.

* `MyClass::MyClass(&x, &source)`: LHS is a variable
* `n$1 =* &__return_param; MyClass::MyClass(n$1, &source)`: LHS is a pointer, which is not supported.

2: There should be a way to pass the fact that "the return value was copy constructed from a source".

Solution:

For 1: This diff matches only when the LHS of the copy constructor is the return variable in SIL, i.e. `__return_param`. Note that it does not support any pointers as LHS in general, because which may require more complicated alias info to be correct.

For 2: This diff introduces a new pulse attribute `CopiedReturn` to pass the fact that "the return value is copy constructed from a source". Then at callsite, it actually modifies the non-disjunctive value with that information as usual.

Reviewed By: ggwg

Differential Revision: D39170753

fbshipit-source-id: f358459a82385b71114223d5522ebf146d2a6a32