[pulse][2/4] add tableau facts to linear arithmetc domain
Summary:
There are two linear arithmetic domains in PulseFormula: linear_eqs and
tableau. Tableau is used to solve inequalities and only deals with
"restricted" (>= 0) variables, and linear_eqs used to deal with
everything else. But, hiding tableau linear equalities from the rest of
the linear equalities introduced incompleteness as nothing was trying to
combine both sets of linear equalities.

Worse, the tableau doesn't care about facts like `a1 = a2 + 1` or even
`a1 = 0`, because they are tautologies as far as the tableau is
concerned, so we were sometimes dropping equalities on the ground.

Fixes https://github.com/facebook/infer/issues/1676

Reviewed By: skcho

Differential Revision: D39310642

fbshipit-source-id: ed73285539647b89fa31b5ea1fde057cdb2e75d9