[IR][Textual] Let propagation transformation
Summary:
Textual allows the programmer to introduce `Let` instruction that put any expression in a logical variable (`Ident.t`) why SIL only allows that for function calls. This transformation remove `id = exp` instructions when `exp` is not a call, using rewritings.

We assume we are in SSA form. Then `id` is defined once and its definition point dominates every usage point. We can then safely replace `id` by `exp` everywhere. But because we have a set of rewriting rules to apply, we try to optimize a bit the rewriting by doing a single pass on the procedure. We have to take care to use the rewriting rules in the right order.

Reviewed By: artempyanykh

Differential Revision: D39308793

fbshipit-source-id: c0400ede74338d3d990b8449e1c5af577b4052ae