[IR][Textual] Out-of-SSA transformation
Summary: This transformation removes SSA arguments in nodes. This is a well known transformation in the SSA literature. We have to generate assignments before jumping to these nodes. The usual subtle points in the literature are so-called critical edges (when an edge starts from a branching and ends on a junction), and the scheduling of parallel assignments. However, these problems are not relevant here because the assignments we generate are stores on variables (then our parallel assignments are sequential too) and this memory locations are dead after the prelude we insert.

Reviewed By: artempyanykh

Differential Revision: D39349557

fbshipit-source-id: 4ba5adb93dab36947dbf50802d45eda57f273863