[infer] Ignore translation unit when comparing global variables
Summary:
This diff ignores translation unit when comparing global variables.

A global variable includes a translation unit field in it.  The
problem is when a global variable is defined in a header file and it
is used in multiple c files, the same global variables may have
different translation units, i.e. they are addressed as different
global variables.

Reviewed By: fantazio

Differential Revision: D39511753

fbshipit-source-id: 34564d84b245bb66e8fd16768bff3e57aa811d14