Define an actually empty Empty domain
Summary:
Until now the so-called "Empty" domain was actually a singleton domain. We define a new empty domain
that has no element, leveraging an empty ocaml type and unreachable patterns. The singleton domain is
renamed into Unit.

This domain may obviously not be used alone and is to be lifted for instance, with WithBottomTop to
define a "boolean" domain.

Note that there was only one occurrence of Empty in the code at
`infer/src/checkers/DisjunctiveDemo.ml`, which seems fine with the new Empty one.

Reviewed By: jvillard

Differential Revision: D39692860

fbshipit-source-id: 139e036d11c52c7f259b238766722f582281ffb8