[pulse][tests] bump disjuncts limit for some erlang tests
Summary:
Some tests were fn_ just because of the disjunct limit. Bumping the
limit also reveals additional FPs.

Looking at a few of these, it seems that the disjunct limit is hit
"artificially": during the analysis we have lots of disjuncts but most
of them are actually unsatisfiable. We only realise this when
normalising at the end, and we get summaries with just a few disjuncts
in the end. This suggests we should do normalisation at intermediate
points or better yet do it incrementally as we go.

Differential Revision: D39729228

fbshipit-source-id: b210d2d376e88d83e958fcf72d20b88a18ecd59e