[pulse] communicate new intervals facts to the rest of the formula
Summary:
The interval domain is only good at one thing: it treats values as
integers (actually, it's probably a source of false negatives when
values are *not* integers! TODO...). Propagate deductions it might make
to the rest of the domains, eg when `x > 0 && x < 4 && x ≠ 2` then only
the intervals will detect that `x = 3`.

Reviewed By: skcho

Differential Revision: D39661802

fbshipit-source-id: 64930c17130c8ec221d8a0af603b9058b030f1e7