Add support for updating of procedure table during analysis
Summary:
The procedure table can be updated during analysis.

In order for the CaptureDatabase to be read-only in analysis phase, we need to add the procedure table into the AnalysisDatabase, and change the update_attributes function to write to it (if we are in the analysis phase).

Finally, we need to potentially query both databases for procedure data.

Reviewed By: ngorogiannis

Differential Revision: D39683647

fbshipit-source-id: e422097967e58f29656dbc0a2d1899abcf35372f