Prototype of a shape pre-analyser for Lineage
Summary:
This prototype aims at inferring all possible fields that may be accessed from every (SIL) variable
of a function being analysed. Albeit being flow-insensitive, this is implemented with the Analyzer
framework to ease possible future extensions (and also because it's no more complex than folding over
the set of instructions of a function).

This does not properly support function calls yet (this will require processing the summary to infer
a relation between the shape of the arguments and the shape of the result).

Currently, we simply output the raw result in the logs. The eventual goal is to use it within the
Lineage analysis.

Reviewed By: rgrig

Differential Revision: D39697295

fbshipit-source-id: 89dadf2016127fd983d82ebbbafda0fc2a51716e