Shape analysis: remove ("garbage-collect") old shape fields on unification
Summary:
Shape unification of shapes s1 and s2, currently:

1. Computes the new set of fields for the final shape.
2. Union_find-unifies the shapes themselves. The id of the unified shape is either s1.id or s2.id.
3. Adds the new fields to the unified shape.

If s2 had fields of its own, and the id of the unified shape happens to be s1.id, then the fields
from s2 will be added as fields of s1.id-as-the-unified-shape-id, but they will also still remain as
fields for an id s2.id that will not be used anymore (as, due to unification, this is no longer the
value associated to any union_find class).

This diff fixes it by removing altogether all the fields of s1.id and s2.id before setting the fields
of the unified shape. Therefore the shape id that happens to become garbage will no longer take some
space in the shape_fields environment.

Reviewed By: rgrig

Differential Revision: D40143808

fbshipit-source-id: 082528765ea73b39b02bf6fdf36f7c37e39b2574