[topl] Simple evaluation of reachability predicate
Summary:
We evaluate reachability predicates X~~>Y (which can occur in guards) by
(1) normalizing the heap, and
(2) doing a depth first search in it

This checks if the pulse state together with the other constraints in
the Topl state *imply* the reachability predicates, which is a stronger
condition than checking that reachability is consistent with the other
known facts. The reasons for this are (a) simplicity of implementation,
and (b) lack of an actual example where it makes a difference.

For (b), the only example I can think of would be a C program in which a
constant is dereferenced and it happens that the constant is the address
of an existing allocated object. But, that would be caught by other
checks.

The other limitation, smaller by comparison, is that we always evaluate
things like 1~~>Y (i.e., with constants) to false. That's probably OK,
but we may want to disallow the possibility of writing such predicates
in the first place.

Differential Revision: D39774010

fbshipit-source-id: 65910f92395b6838c5c68a5dddf176251adb3fdb