[logs] log results too
Summary:
It's convenient to have the results in the logs too in addition to the
console.

They are printed as a single string so we don't get the usual prefixes
of `[PID][category]` before each line. It doesn't seem worth the effort
to "fix" that and make copy/paste harder in the process.

Reviewed By: ngorogiannis

Differential Revision: D40260711

fbshipit-source-id: fa6690b4d4fced0e458b7f8be2327eef33625ca0