Make the Shape analysis interprocedural
Summary:
We define a Summary for the shape analysis, that is essentially a frozen version of the shaping
environment of the function (as we will need the full information later for Lineage). This summary is
used when calling functions, by fetching only the part relevant to the parameters and the return
value.

Reviewed By: rgrig

Differential Revision: D40143805

fbshipit-source-id: ab0d74407b1c5f1562986afbf704d3215834b7d7