Sanity check for new_eqs
Summary:
Updating formulas (e.g., normalizing) produces "new_eqs" that are
supposed to be applied to other parts of the abstract state (such as the
heap). These new_eqs are conceptually equalities, but the implementation
tries to maintain the invariant that applying them as substitution, in
order, preserves all information.

Example 1: If we treat x=y, y=z as the substitutions x->y, y->z, applied
in the left-to-right order, then all x,y,z become z.
That's OK.

Example 2: If we treat y=z, x=y as substitutions y->z, x->y applied in
the left-to-right order, then we are left with both z and y.
That's not OK.

This check is expensive (non-constant; quadratic in the size of new_eqs),
so it is only turned on by `--pulse-sanity-checks`.

Reviewed By: jvillard

Differential Revision: D40301172

fbshipit-source-id: 93dea56a9adf53e0fa7904cecab2bc04a0fc8183