[pulse] Create a view of the CFG where there's no edge from exception sink to exit node
Summary:
We modify the PulseExceptional module so that the `fold_normal_succs` and `fold_normal_preds` functions required by the ProcCfg signature get overridden, such that we create a view of the CFG for Pulse where there is no edge from the exception sink to the exit node.
- `fold_normal_succs` filters out the exit node from the CFG successors (when the node is the exceptions sink).
- `fold_normal_preds` filters out the exceptions sink from the CFG predecessors (when the node is the exit node).

We want to extract the postconditions of the exceptions sink node.
* We replace the call to `DisjunctiveAnalyzer.compute_post` with (i) construction of the invariant map `inv_map` (from which the postconditions of any node can be retrieved), (ii) calls to `DisjunctiveAnalyzer.extract_post` to extract the postconditions of both the exit node, and the exceptions sink. This ensures we avoid duplicating the work associated with constructing the invariant map.

Note that we currently find the exceptions sink by searching the list of nodes in the procedure's CFG (which we later plan to refactor, by modifying the relevant frontends to include the exceptions sink as a field of the `ProcDesc`). Also, we currently assume that there is at most one exceptions sink node in a procedure.

The exception sink’s postconditions get extracted, and put into a list `states_at_exn_sink : ExecutionDomain.t list`. These ExecutionDomains are also provided to `PulseSummary.of_posts`.

However, note that at this stage, the contents of `states_at_exn_sink` are all ContinueProgram due to the behaviour of exceptional control-flow edges. Therefore, this diff has no actual effect on the final summaries that are generated for procedures (i.e. the summaries generated for procedures will be the same as if the edge from the exception sink to the exit node were present). A later diff will turn the postconditions of the exceptions sink into ones of type `ExceptionRaised`.

Differential Revision: D39258826

fbshipit-source-id: 2344b25bf88a01f92fbdb29a848bd837678f5771