[pulse] Refactor exec_summary_of_post_common so that we can reuse it for ExceptionRaised execution domains
Summary:
We refactor `exec_summary_of_post_common` in PulseSummary.ml so that (in a later diff) this logic can be reused for `ExceptionRaised` states.

The transformations applied in the `ContinueProgram` case of `exec_summary_of_post_common` are extracted into a function called `summarise`, that takes a parameter `~exec_domain_of_summary`. In the `ContinueProgram` case we simply forward the `continue_program` argument of `exec_summary_of_post_common`.

Reviewed By: rgrig

Differential Revision: D39262518

fbshipit-source-id: 9f3d9db84ae660eb50afac36c98206f9df26b1e3