[topl] Simple evaluation of field access
Summary:
Evaluate field access in Topl. To access a field with name `f` on an abstract value `v1`, we use the following edges: `v1 --f--> v2 --*--> v3`, where `v3` is the final abstract value that we are looking for (that is, we do both the field access and then the dereferencing).

If any of the abstract values/edges could not be found in the heap, we ignore this transition in Topl (to avoid FPs). To be able to "ignore", we had to do some modifications to the `predicate` and `Constraint` types in Topl to facilitate the `False` predicate.

Reviewed By: rgrig

Differential Revision: D40026565

fbshipit-source-id: c4e654be2553389f9840dbbf0fc6b9ea052e8044