[pulse][taint] copy taints from values to their content
Summary:
When tainting a structure or a pointer, we also taint their content.
Howver, the way we do it relies on what already exists in memory. This
means that if we never dereferenced the pointer or accessed a field of
the structure before the tainting, then they will not be tainted when we
try to access them later. This leads to FN as the now wroking test
shown.

The fix is to taint the content the first time it is accessed, like we
would have done it if it had been accessed before the containing value
was tainted

Reviewed By: dulmarod

Differential Revision: D40350976

fbshipit-source-id: 6f72b12ae47cc947070e4bcf74f366c725de0ce7