[FCP] Extend Location with macro file/line for macro expansions
Summary:
Currently, Infer reports issues at macro usages (call-sites) but this is pretty confusing for developers because macros are preprocessed and their definitions are expanded in the callsite. However, when developer looks at the call-site they cannot see any of the code being reported on. The aim of this diff is to improve that by extending the location information with file and actual line number inside the macro definition so that checkers can decide to
- include that in the trace (see next diff on the stack)
- tweak the messaging of the issues on macro locations

To include such information on `Location.ml` we need to export macro file+line number from clang plugin. We do this similar to `SourceLocation` so that we don't duplicate identical line numbers/files. While doing that, we have to also keep track of `is_macro` because the macro files might not be the same, so we should only mutate/re-use for things that are macro locations.

Reviewed By: jvillard

Differential Revision: D40105645

fbshipit-source-id: 2f517e4168ec699dac3393d0c08d6cc3de0093be