[pulse][taint] Record a false positive that implies internal pointer tainting
Summary:
Because variables and fields are manipulated as pointers in Pulse and
when tainting a value we also taint all values accessible from it, we
taint structures' internal pointers to fields. As a result, if we
reassign a field to an untainted value, and later sink the address of
that field, we get a FP because that pointer was never changed so its
taint remained although it only existed because of the way we do things,
not because the user intended it.
We should avoid tainting internal pointers unless the code explicitly
manipulate said pointers:
`taint(s)` should taint `s` and `s.f` but not `&(s.f)` (the internal
pointer)
`taint(&(s.f))` should taint `&(s.f.)` and `s.f`

Reviewed By: skcho

Differential Revision: D40175909

fbshipit-source-id: 8de7add25009bf7454f815b17bd1b28cf56d3a8e