[erl-frontend] Fix locations for included files
Summary: When we include files, one AST can include multiple `file` attributes, and the semantics is that all following line/column numbers should correspond to this file, until we have an other `file` attribute. Previously, we went through these attributes during pre-processing and overridden the location in the environment that we initialized. This means we always had the path from last `file` attribute. In this diff, we shift this processing to translation, where we always replace the current path in the environment (and consequently, translate other forms with the correct path).

Reviewed By: rgrig

Differential Revision: D40511212

fbshipit-source-id: 870d12d102a77baa88d1ec94592c8074eaa3b9e0