[clang] emit correct number of record items for locations
Summary:
Most of the time a location did not come from macro expansion so we can
save on the number of fields we emit. Previously we would still emit at
least one field for `is_macro` even when it's `false` (the default),
thus forcing the biniou backend to emit a dummy field.

Reviewed By: dulmarod

Differential Revision: D40559191

fbshipit-source-id: d75cd4684f06d728a847b7da40c10971778e92db