[db][7/6] load payloads eagerly during procedures analysis
Summary:
This seems to help with perf overall, although the variability is high.
This basically restores the previous behaviour of eagerly loading
payloads, but only for callees of functions, when we are very likely to
read those payloads back. This avoids the cost of making one DB lookup
per payload in that case. The file analysis still loads payloads lazily
so the mem usage remains the same in tests.

Reviewed By: ezgicicek

Differential Revision: D40717206

fbshipit-source-id: b087a0c5d370f1f8e9bdac499a6b210f293a373d