[unnecessary copy] Don't track copy assignment from member fields
Summary: We stop tracking copy assigned source objects that are originating from member fields. This is because suggesting moves from such source objects is not safe as other procedures might access the member field and without a whole program analysis, it is not possble for us to detect such accesses.

Reviewed By: skcho

Differential Revision: D40858642

fbshipit-source-id: 5504c1a8a1e2293b94e4898ea3a91a0f037ad2e9